Update ex13.py
Was supposed to combine raw_input with a argv to make a script that gets more input from a user. 

Not sure if this suffices, since I have to put the parameters in first when I call up the program in the command line, but then  users can enter in anything they want when prompted by the raw_input section.